requirements.txt
discord.py
pytz
python-dotenv
quart
aiosqlite
hypercorn
pytest
pytest-asyncio
